refactor: split shipping form into multiple components
